Move richards-wasm to JS-style scoring
