+++ wasm/richards/build.log
@@ -1,0 +1,5 @@
+Built on 2025-01-16T15:10:49Z
+Toolchain versions
+emcc (Emscripten gcc/clang-like replacement + linker emulating GNU ld) 3.1.73 (ac676d5e437525d15df5fd46bc2c208ec6d376a3)
+Building...
+Building done
